rewrite check_dhcpd-pools
Before this commit, we expected every pool to have its own /8.
This assumption was never true. Now NMS Prime puts a comment in
the DHCPD shared-network config, such that we can easily retrieve
the start and end IP address of a pool and for which type
(CM, MTA, CPEPriv, CPEPub) it is used.